backend/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Alternative: FastAPI (uncomment if using FastAPI instead)
# fastapi==0.104.1
# uvicorn==0.24.0

# PubMed API Integration
biopython==1.81
requests==2.31.0

# Text Processing & Embeddings
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.2.0

# Vector Database
chromadb>=0.4.15

# Alternative vector databases (uncomment if using)
# pinecone-client==2.2.4
# faiss-cpu==1.7.4

# LLM APIs
openai>=1.3.0
# anthropic==0.7.8  # Uncomment if using Anthropic

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Configuration & Environment
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Data Validation
pydantic==2.5.0

# HTTP Client
httpx==0.25.1

# Text Processing
nltk==3.8.1

# Development & Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0